accrual type fix

+++ internal/accrual/client.go
@@ -16,9 +16,9 @@ type Client struct {
 }
 
 type OrderInfo struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual *int64 `json:"accrual,omitempty"`
+	Order   string   `json:"order"`
+	Status  string   `json:"status"`
+	Accrual *float64 `json:"accrual,omitempty"`
 }
 
 var (
